Fixes payment screen tests and some amount input tests
